Combine bulk check func into one controller
