--- CheckYourEligibility.Admin/Views/Shared/BulkCheckUpload.cshtml
@@ -1,110 +1,0 @@
-@model CheckYourEligibility.Admin.ViewModels.BulkCheckUploadViewModel
-
-@{
-    ViewData["Title"] = "Run a batch check for multiple parents or guardians";
-    var errorMessage = TempData["ErrorMessage"] as string;
-}
-
-<div class="govuk-grid-column-full">
-    <a class="govuk-back-link backLinkJS" href="@Url.Action("Index", "Home")">Back</a>
-
-    @if (!string.IsNullOrEmpty(errorMessage))
-    {
-        ViewData["Title"] = "Error: Check Details";
-        <div id="error-summary" class="govuk-error-summary" data-module="govuk-error-summary">
-            <div role="alert">
-                <h2 class="govuk-error-summary__title">There is a problem</h2>
-                <div class="govuk-error-summary__body">
-                    <ul class="govuk-list govuk-error-summary__list">
-                        <li>
-                            <a class="govuk-error-message" href="#fileUploadDropZone">@errorMessage</a>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-        </div>
-    }
-    <div class="moj-page-header-actions">
-        <div class="moj-page-header-actions__title">
-            <span class="govuk-caption-l">@ViewData["Title"]</span>
-            <h1 class="govuk-heading-l">Run a manual batch check</h1>
-        </div>
-
-        <div class="moj-page-header-actions__actions">
-            <div class="moj-button-group moj-button-group--inline">
-                <a href="@Url.Action(Model.HistoryAction, Model.HistoryController)" class="govuk-button govuk-button--secondary" data-module="govuk-button">
-                    Batch checks history
-                </a>
-            </div>
-        </div>
-
-    </div>
-
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">
-            <h2 class="govuk-heading-m">Upload our CSV template</h2>
-
-            <ol class="govuk-list govuk-list--number">
-                <li>
-                    Download the <a href="@Url.Action(Model.DownloadTemplateAction, Model.DownloadTemplateController)" class="govuk-link">batch check template (CSV, 1KB)</a>
-                </li>
-                <li>
-                    @(Model.isEnhanced ? 
-                    "Add the details of the parents or guardians you need to check, as well as the children who will be added to the application. Add a row for each child" :
-                    "Add the details of the parents or guardians you need to check.")
-                </li>
-                <li>
-                    Make sure you include:
-                    <ul class="govuk-list govuk-list--bullet">
-                        @foreach (var guidanceItem in Model.GuidanceItems)                        
-                        {
-                            <li>@guidanceItem</li>
-                        }
-                        @if (Model.isEnhanced && !Model.isSchool)
-                        {
-                            <li>URN of the school the child attends on a full-time basis</li>
-                        }
-                        
-                    </ul>
-                </li>
-                <li> Save it as a CSV file.</li>
-            </ol>
-            <div class="govuk-warning-text">
-                <span class="govuk-warning-text__icon" aria-hidden="true">!</span>
-                <strong class="govuk-warning-text__text">
-                    <span class="govuk-visually-hidden">Warning</span>
-                    Your saved file must match the template fields and contain fewer than 5000 rows.
-                </strong>
-            </div>
-            <hr class="govuk-section-break govuk-section-break--m govuk-section-break hideSectionBreak">
-
-            <h3 class="govuk-heading-m">Upload a CSV file</h3>
-        </div>
-    </div>
-    <form asp-controller="@Model.FormController" asp-action="@Model.FormAction" method="post" enctype="multipart/form-data">
-        <div id="fileUploadField" class="govuk-form-group @(string.IsNullOrEmpty(errorMessage) ? "" : "govuk-form-group--error")">
-            @if (!string.IsNullOrEmpty(errorMessage))
-            {
-                <p id="file-upload-1-error" class="govuk-error-message">
-                    <span class="govuk-visually-hidden">Error:</span> @errorMessage
-                </p>
-            }
-            <div class="govuk-form-group">
-                <label class="govuk-label" for="fileUploadDropZone">
-                    Upload a file
-                </label>
-                <div class="govuk-drop-zone" data-module="govuk-file-upload">
-                    <input class="govuk-file-upload @(string.IsNullOrEmpty(errorMessage) ? "" : "govuk-file-upload--error")"
-                           id="fileUploadDropZone" name="fileUpload" type="file" accept=".csv"
-                           aria-describedby="file-upload-1-error" />
-                </div>
-            </div>
-        </div>
-
-        <button class="govuk-button" data-module="govuk-button">
-            @Model.SubmitButtonText
-        </button>
-    </form>
-</div>
-
-<script src="/js/validationSummary.js"></script>
